Apply clang-tidy to folder source
All the code in folder source is checked with clang-tidy. These issues
have been found and corrected:
- explicit cast to unsigned long required when using long format
  specifier
- static_cast must not be used
- last argument of va-args must not use standard promotion.
- clang-tidy must be called with compiler option --std=gnu99 since
  arm-none-eabi does not fully support c11.

+++ source/app/Presentation.c
@@ -395,7 +395,7 @@ static void HandleUnrecoverableError(uint32_t errorCode) {
   Screen_DisplayFourDigits(errorCode, rowBottom, Screen_DisplayMinusBottom);
   Screen_UpdatePendingRequests();
   LOG_ERROR("Unrecoverable error %lu!!\nThe system needs to be rebooted\n!",
-            errorCode);
+            (unsigned long)errorCode);
 }
 
 static void LogFirmwareVersion(void) {
